Move ItemEntityMockTest's dropItem tests to WorldMockTest
